+++ .changeset/config.json
@@ -7,5 +7,5 @@
   "access": "public",
   "baseBranch": "main",
   "updateInternalDependencies": "patch",
-  "ignore": ["apm_tools", "configs", "scripts", "docs"]
+  "ignore": ["apm_dependencies", "configs", "scripts", "docs"]
 }
