+++ apm_dependencies/core/index.ts
