+++ apm_dependencies/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "apm_tools",
+  "name": "apm_dependencies",
   "type": "module",
   "version": "1.0.0",
   "license": "MIT",
